Sensor reading works and in a separate file
