+++ PIC18F27K42/nbproject/Makefile-genesis.properties
@@ -1,5 +1,5 @@
 #
-#Sun Feb 28 00:07:39 EST 2021
+#Sun Feb 28 00:51:31 EST 2021
 proj.dir=D\:\\projects\\xTendLoRa\\LWmesh\\PIC18F27K42
 atbootable.languagetoolchain.version=2.31
 base.languagetoolchain.version=2.31
@@ -11,7 +11,7 @@ user_application.languagetoolchain.dir=C\:\\Program Files\\Microchip\\xc8\\v2.31
 MBRTU.Pack.dfplocation=C\:\\Program Files\\Microchip\\MPLABX\\v5.45\\packs\\Microchip\\PIC18F-K_DFP\\1.4.87
 atbootable.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=d9404904a11cd8baf0bd6c1f62dd8c23
 transparent_mode.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=d9404904a11cd8baf0bd6c1f62dd8c23
-configurations-xml=5975bf0f8466f8480b3fb6db53fab504
+configurations-xml=43ab9b2b57de01566a37cd07afe7dbc3
 atbootable.Pack.dfplocation=C\:\\Program Files\\Microchip\\MPLABX\\v5.45\\packs\\Microchip\\PIC18F-K_DFP\\1.4.87
 base.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=d9404904a11cd8baf0bd6c1f62dd8c23
 host.platform=windows
