+++ PIC18F27K42/nbproject/Makefile-impl.mk
@@ -27,7 +27,7 @@ CLEAN_SUBPROJECTS=${CLEAN_SUBPROJECTS_${SUBPROJECTS}}
 PROJECTNAME=PIC18F27K42
 
 # Active Configuration
-DEFAULTCONF=MBRTU
+DEFAULTCONF=user_application
 CONF=${DEFAULTCONF}
 
 # All Configurations
