+++ app/app_src/application.c
@@ -39,6 +39,9 @@ Copyright 2020 Samuel Ramrajkar
 #ifdef FULLFEATURE
 #include "I2C_EEPROM.h"
 #endif
+#ifdef USERAPP
+#include "user_app.h"
+#endif
 #if (__32MM0256GPM048__)
 #include "uart3.h"
 #include "uart2.h"
@@ -1368,15 +1371,6 @@ static void cmdSendPing(char *cmd){
     printf("OK:%u\r\n", tx_buffer[buf_id].msgid);
 }
 
-/*!
- * \brief process distance data
- * \param [OUT] None.
- * \param [IN] None.
- */
-static uint8_t processDistanceData(uint8_t distance_data[3]){
-    uint16_t distance = distance_data[0] * 256 + distance_data[1];
-    // TODO(anyone): Handle data obtained above
-}
 
 /*!
  * \brief execute AT commands
@@ -2464,132 +2458,6 @@ static void exract_sink_addr(uint8_t* dataptr){
     printf("SINK SET:%02X%02X\r\n",sinkAddr0, sinkAddr1);
 #endif                    
 }
-
-#ifdef USERAPP
-static void user_application(void){
-    //Check if RS485 tx should be turned off
-#if 0
-    if(tx_done && U1ERRIRbits.TXMTIF){
-//       TXEN_SetLow();
-       tx_done = 0;
-    }
-#endif
-    //Check if there is a command halfway thru and it timed out   
-    if(((sensorStateVar != lookingForHeader) || (sensorStateVar != processData)) && 
-            (!get_timer0base(&ATTimeoutTimer)))        {
-        sensorStateVar = resetSensorMachine;
-    }
-    switch(sensorStateVar){
-        case initMessage:            
-            sensorStateVar = resetSensorMachine;
-            break;
-        case lookingForHeader:
-            //Read a byte if it is available
-            #if (_18F27K42 || _18F47K42 || _18F26K42)
-            if(UART1_is_rx_ready()){
-            #endif
-            #if (__32MM0256GPM048__)
-            if(true == UART3_IsRxReady()){
-            #endif  
-                //There is a byte available
-                #if (_18F27K42 || _18F47K42 || _18F26K42)
-                if(UART1_Read() == 0xff)
-                #endif
-                #if (__32MM0256GPM048__)
-                if(UART3_Read() == 0xff)
-                #endif
-                {
-                    //Found 0xff
-                    sensorStateVar = lookingForDataH;
-                    set_timer0base(&ATTimeoutTimer, sensorTimeout);
-                }
-            }                
-            break;        
-        case lookingForDataH:
-            #if (_18F27K42 || _18F47K42 || _18F26K42)
-            if(UART1_is_rx_ready()){
-            #endif
-            #if (__32MM0256GPM048__)
-            if(true == UART3_IsRxReady()){
-            #endif 
-                //There is a byte available
-                #if (_18F27K42 || _18F47K42 || _18F26K42)
-                uint8_t data = UART1_Read();
-                #endif
-                #if (__32MM0256GPM048__)
-                uint8_t data = UART3_Read();
-                #endif
-                //Found DATA_H
-                sensorStateVar = lookingForDataL;
-                distanceData[distanceDataCounter++] = data;
-            }
-            break;            
-        case lookingForDataL:
-            //Found a AT command start now read till it is /r
-            #if (_18F27K42 || _18F47K42 || _18F26K42)
-            if(UART1_is_rx_ready()){
-            #endif
-            #if (__32MM0256GPM048__)
-            if(true == UART3_IsRxReady()){
-            #endif 
-                #if (_18F27K42 || _18F47K42 || _18F26K42)
-                uint8_t data = UART1_Read();
-                #endif  
-                #if (__32MM0256GPM048__)
-                uint8_t data = UART3_Read();
-                #endif
-                // Found DATA_L
-                sensorStateVar = lookingForSum;
-                distanceData[distanceDataCounter++] = data;
-            }
-            break;         
-        case lookingForSum:
-            //Found an end of AT command. Wait for /n
-            #if (_18F27K42 || _18F47K42 || _18F26K42)
-            if(UART1_is_rx_ready()){
-            #endif
-            #if (__32MM0256GPM048__)
-            if(true == UART3_IsRxReady()){
-            #endif 
-                #if (_18F27K42 || _18F47K42 || _18F26K42)
-                uint8_t data = UART1_Read();
-                #endif  
-                #if (__32MM0256GPM048__)
-                uint8_t data = UART3_Read();
-                #endif
-                // Found sum
-                distanceData[distanceDataCounter++] = data;
-        
-                // checksum validation
-                uint8_t checksum = (0xff + 
-                                    distanceData[0] + 
-                                    distanceData[1]) &
-                                    0x00ff;
-                if(checksum == distanceData[2])
-                {
-                    sensorStateVar = processData;
-                }
-            }
-            break;
-            
-        case processData:
-            // TODO(anyone): Add distance calculation logic
-            processDistanceData(distanceData);
-            sensorStateVar = resetATMachine;
-            break;
-        case resetATMachine:
-            sensorStateVar = lookingForHeader;
-            commandByteCounter = 0;
-            //clear the ATcommand buffer
-            memset(distanceData,0,sizeof(distanceData));
-            set_timer0base(&ATTimeoutTimer, sensorTimeout);
-            break;
-            
-        default:
-            sensorStateVar = resetSensorMachine;
-    }
-}
-#endif
 
 inline void application(void){
     start_loop_timer();
